RSVP availabilitly?
Is it possible to make tickets in a way so its "selective seating"? If you have 100 tickets for a graduation, and each seat corresponded to a ticket number, there would only be a limit of one ticket per seat. This way people can buy the tickets for the seats they want. Makes it easy for RSVP events 