Escape unfiltered HTML in the Administration Panels.
An Administrator with the unfiltered_html capability can cause XSS within the admin panel through some form fields. Core's policy is to display the unfiltered HTML raw on the front-end, but escape it on the back-end. This commit brings CampTix in line with Core's treatment of the HTML.

As a security concern, this is very minor, since it requires a Administrator account, so this is being done more as a precaution against extreme edge cases than anything else.