Revert "Allow admins to edit other user's tickets."
This reverts commit f7815f12689aa376ac0cf92b4408086045d8fa3f.

When dealing with unconfirmed tickets, there's no way of knowing if the ticket is intended for the admin who is editing it, or for another user. So, in that situation, we have to either always set the username, or never set it.

If we always set it, that works fine when the ticket is intended for the admin, but produces undesired behavior when the ticket is intended for someone else. If we never set the username, that works fine when the ticket is intended for someone else, but fails when it's intended for the admin.

For example, if an admin is editing a ticket intended for someone else and we set their username, then the ticket would be considered confirmed, and the actual ticket owner would not get e-mail reminders asking them to confirm the ticket. It would also affect any external actions hooked to `camptix_rl_registration_confirmed`.

Or in the reverse situation, if they're editing their own ticket, their username would never be saved and the ticket would be perpetually unconfirmed, resulting in unending e-mail confirmation reminders.

So it's better to just prevent admins from editing other user's tickets in the first place. They can still edit the `tix_attendee` post from the Administration Panels if they really want to edit the attendee.