Strip slashes from `$_POST` to avoid outputting slashed strings.
Leaving the slashes leads to bugs where names with apostrophes are output with slashes during a checkout error, or users are blocked from logging in to `[camptix_private]` pages because the slash causes the string comparison to fail.

The values will be escaped with a more appropriate function like `esc_attr()` or `esc_html()` when they are output.