Shortcodes: Remove unused and unsafe form action URL.
This wasn't being echo'd, so it wasn't actually doing anything, but letting `add_query_arg()` fall back to using `$_SERVER['REQUEST_URI']` can introduce malicious user input.

See https://make.wordpress.org/plugins/2015/04/20/fixing-add_query_arg-and-remove_query_arg-usage/

It's being removed because it's dead code and also to avoid the risk of introducing a vulnerability in the future if a developer noticed it wasn't being echo'd and fixed that, but didn't notice the security issue.