Raise posts_per_page limit in get_tickets_post_id() to 50.
A value of 5 assumes that there won't be more than 5 pages that use any of the [camptix*] shortcodes, but it's not unreasonable for a site to have more than 5. If they do, then the [camptix] shortcode might not be found, and get_tickets_url() will generate broke URLs like:

https://example.org/tix_attendee/jane-doe/?tix_action=edit_attendee&tix_attendee_id=49&tix_edit_token=85db5efa526b08980a303ec2c2615bf5#tix

The value was initially set to 1 in 880ee42c398d7e65d9b5b339224bece8de4c33f6, presumably for performance reasons. It was later updated to 5 in a2e41f59c01e27136e0c6add88e0a74d63c898ed, presumably because this bug was encountered, and 5 was assumed to be a safe value.

There's no indication that this function is not performant in the real world, and setting a limit introduces bugs at unpredictable times, so it's premature optimization. If we do discover performance issues in the future, we can just cache the result instead of limiting the number of posts.