Set Reply-To header instead of From, to avoid being flagged as spam.
I'm assuming that the intention of the original code was to set the name part of the From header, so that attendees wouldn't be confused by receiving an e-mail from "WordPress" instead of the event; and perhaps also so that attendees could reply to the message and reach the organizer, rather than having it bounce because wordpress@domain isn't a valid address.

Those are both good goals, but a serious problem was introduced because the address part of the header was also overriden. If the domain of the admin e-mail address doesn't match the domain of the web server, then the receiving MTA will often severly penalize the message when it scans it for spam, because it will look like we're trying to pose as another domain.

Both goals can both be achieved in a better way, but only setting the name part of the From header, and using a Reply-To header.