Fix checkboxes with multiple options and ampersands
When a checkbox-type question is saved multiple times, the values
in the database containing an ampersand are converted to &amp; and the
Edit Information UI comparison no longer works. The escaping is a
separate bug, but to fix the UI for existing data let's compare to
an array of escaped values in addition to the ones originally stored.

Fixes #144