Hide the overlapping "Attended" column title in mobile views
In the mobile view (<782px) of the attendee list, the Attended button is overlapped by a "Attended" label which comes from the list table. Although I'm sure the label serves a purpose, it causes problems accessing the button (and just looks ugly.)

See https://cloudup.com/caiDk82n2au for a before and after.