Attendees shortcode: Lazy load attendee avatars
A large attendee list causes a large number of concurrent requests to
be sent from the browser to Gravatar, which bogs down the page loading
performance.

This inserts a placeholder element for all but the first 50 attendee
avatars in a list. A jQuery plugin called Appear is used to detect when
a placeholder element is about to scroll into view, which triggers an
Ajax call to retrieve the actual markup for that attendee's avatar.
Jetpack's jquery.spin script is used to provide a visual indicator that
an avatar is in the process of loading.

Refs https://meta.trac.wordpress.org/ticket/1760