Attendees shortcode: Refactor lazy load to use JS template
Instead of retrieving the HTML markup for each avatar via Ajax, the
needed avatar element attributes are added to the placeholder as data
attributes, and then the placeholder is replaced with a JS template
when it comes into view, and the avatar image gets loaded.

Thanks to @iandunn for the idea. This is WAY faster.