Attendees shortcode: Lazy load attendee avatars
A large attendee list causes a large number of concurrent requests to
be sent from the browser to Gravatar, which bogs down the page loading
performance.

This inserts a placeholder element for all the attendee avatars when
rendering the HTML markup. A jQuery plugin called Appear is then used to
detect when a placeholder element is about to scroll into view, which
triggers an event that replaces the placeholder element with an img tag,
using a JS template. Jetpack's jquery.spin script is used to provide a
visual indicator that an avatar is in the process of loading.

Refs https://meta.trac.wordpress.org/ticket/1760