Attendees shortcode: Move `camptix_attendees_shortcode_init` action
This attempts to resolve the following error:

  PHP Warning: Invalid argument supplied for foreach() in
  /home/wordcamp/public_html/wp-content/plugins/camptix/
  addons/field-twitter.php on line 36

When the shortcode functionality was refactored to change the way it
cached the attendees list content, a regression was introduced because
there were some cases where the `camptix_attendees_shortcode_init`
action was not called during the markup rendering, which caused some
other addon code to throw errors like the above. This ensures the
action is always fired during content rendering.