Stripe: Fix refund bug, improve logging
The idempotency key was not working when submitting refund requests
because we were using the same key that was used during the original
charge transaction. Now we're generating a new key for the refund
request.

This also links logging related to charge and refund errors to their
relevant attendee posts.