Stripe: Change handling of numbers in fractional unit calculation
The base amount was being cast as an integer before being combined with
the multiplier. This mean that, e.g., the cents would be stripped off
of a dollar amount. This commit changes it so that the base amount is
first cast as a float, to keep the cents, and then cast as an integer
after the fractional unit amount has been calculated.