Checkout: Fix validation/error message for payment method
The logic to determine whether or not to add an error flag for an
invalid payment method was based on checking the existence and contents
of `order['price']`. However, that array key doesn't exist, because
the price is listed for each ticket in the `items` sub-array. Instead
we need to check the existence and contents of `order['total']`.