Memoize buyer data to save SQL queries to figure out ticket buyer of an attendee.
Attendee data export was timing out for 2018.us.wordcamp.org. Apparently the bottleneck was figuring out the buyer data. We were making meta queries to figure out the attendee buyer data resulting in an `n+1` query pattern. Given how we insert buyer data, we can just treat first attendee with a different access_token, to be the ticker buyer for rest of the attendees of with the same access token.