Attendees Shortcode: Don't key attendee IDs with email
For some reason the render callback was taking an array of attendee
IDs and re-keying it to the attendee's email address. In the case
where a person purchased multiple tickets and didn't specify a separate
email address for each ticket, this meant that often the
attendee IDs associated with that email address would keep getting
overwritten, so that only one attendee per email address would be
output in the list. (This was further complicated by the fact that the
query is paginated, so for events with more than 200 attendees there
was the possibility that some attendees wouldn't get overwritten if
they were on separate pages). The email address key isn't actually
used for anything, so by removing this step, we ensure that all
attendees with the same email are able to be output in the list.

I suspect this also addresses #194.